% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inpi_codeRole}
\alias{inpi_codeRole}
\title{Codes role de l'Inpi}
\format{
A data frame with x rows and y variables:
\describe{
\item{code}{code role inpi}
\item{libelle}{libelle role inpi}
}
}
\usage{
inpi_codeRole
}
\description{
Dataset contenant les codes role et leur libelle
Les variables sont les suivantes
}
\keyword{datasets}
