% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_acte_infos}
\alias{inpi_get_acte_infos}
\title{inpi_get_acte_infos}
\usage{
inpi_get_acte_infos(siren = "", simp = TRUE)
}
\arguments{
\item{siren}{le siren a interroger}

\item{simp}{simplifier le json ? default = TRUE}
}
\value{
json non parse de la reponse a la requete
}
\description{
retourne le json de reponse de la requete sur un acte
}
\examples{
\dontrun{
inpi_get_acte_infos(siren = "")
}
}
