% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_acte_pdf}
\alias{inpi_get_acte_pdf}
\title{inpi_get_acte_pdf}
\usage{
inpi_get_acte_pdf(id = "")
}
\arguments{
\item{id}{l'identifiant a interroger}
}
\value{
le fichier pdf de l'acte
}
\description{
retourne le pdf de l'acte demande
}
\examples{
\dontrun{
inpi_get_acte_pdf(id = "")
}
}
