% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_bilan_pdf}
\alias{inpi_get_bilan_pdf}
\title{inpi_get_bilan_pdf}
\usage{
inpi_get_bilan_pdf(id = "")
}
\arguments{
\item{id}{l'identifiant du bilan a consulter}
}
\value{
le fichier pdf du bilan
}
\description{
retourne le pdf du bilan demande
}
\examples{
inpi_get_bilan_pdf(id = "")
}
