% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_bilan_saisis}
\alias{inpi_get_bilan_saisis}
\title{inpi_get_bilan_saisis}
\usage{
inpi_get_bilan_saisis(id = "", simp = TRUE)
}
\arguments{
\item{id}{l'identifiant a interroger}

\item{simp}{simplifier le json ? default = TRUE}
}
\value{
json non parse de la reponse a la requete
}
\description{
retourne le json de reponse de la requete sur un bilan saisis
}
\examples{
inpi_get_bilan_saisis(id = "")
}
