% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_by_acti}
\alias{inpi_get_by_acti}
\title{inpi_get_by_acti}
\usage{
inpi_get_by_acti(activite = "COMMERCIALE", simp = TRUE)
}
\arguments{
\item{activite}{le nom de l'activite a interroger}

\item{simp}{if TRUE (default) simplifyVector}
}
\value{
json non parse de la reponse a la requete sur activite
}
\description{
retourne le json de reponse de la requete sur un secteur d'activite
}
\examples{
\dontrun{
inpi_get_by_acti(activite = "ARTISANALE", simp = TRUE)
}
}
