% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_by_name}
\alias{inpi_get_by_name}
\title{inpi_get_by_name}
\usage{
inpi_get_by_name(nom = "", simp = TRUE)
}
\arguments{
\item{nom}{le nom a chercher}

\item{simp}{simplifier le json ? default = TRUE}
}
\value{
json non parse de la reponse a la requete sur nom
}
\description{
retourne le json de reponse de la requete sur un nom d'entreprise (requete type content)
}
\examples{
inpi_get_by_name(nom = "FIDANI")
}
