% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_extrait}
\alias{inpi_get_extrait}
\title{inpi_get_extrait}
\usage{
inpi_get_extrait(siren, url = FALSE)
}
\arguments{
\item{siren}{le siren a interroger}

\item{url}{si true la fonction ne retourne que le lien, pas le doc}
}
\value{
un pdf de l'extrait inpi ou un lien
}
\description{
retourne l'extrait inpi du siren demande
}
\examples{
\dontrun{
inpi_get_extrait(siren = "123456789", url = FALSE)
}
}
