% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_sirens}
\alias{inpi_get_sirens}
\title{inpi_get_sirens}
\usage{
inpi_get_sirens(listesiren = c("", ""), simp = TRUE)
}
\arguments{
\item{listesiren}{un vecteur contenant les sirens a interroger}

\item{simp}{simplifier le json ? default = TRUE}
}
\value{
json non parse de la reponse a la requete sur plusieurs sirens
}
\description{
retourne le json de reponse de la requete sur plusieurs sirens
}
\examples{
inpi_get_sirens(listesiren = c("123456789", "987654321"))
}
