% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_token}
\alias{inpi_get_token}
\title{inpi_get_token}
\usage{
inpi_get_token(user = "gillesfidani@gmail.com", varenv = "INPI_CODE")
}
\arguments{
\item{user}{le mail renseigne lors de l'inscription}

\item{varenv}{le nom de la variable d'environnement stockant le mot de passe}
}
\value{
string le token
}
\description{
retourne le token necessaire aux requetes de l'API
}
\examples{
inpi_get_token(user = "gillesfidani@gmail.com", varenv = "INPI_CODE")
}
