% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_histo}
\alias{inpi_histo}
\title{inpi_histo}
\usage{
inpi_histo(x)
}
\arguments{
\item{x}{la reponse json a parser}
}
\value{
data.frame avec les infos sur l'historique
}
\description{
retourne les colonnes de l'historique
}
\examples{
inpi_histo(x = resultat1)
}
