% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_meta_px}
\alias{inpi_meta_px}
\title{inpi_meta_px}
\usage{
inpi_meta_px(x)
}
\arguments{
\item{x}{la reponse json a parser}
}
\value{
data.frame avec les infos sur les PX
}
\description{
retourne les colonnes des px
}
\examples{
\dontrun{
inpi_meta_px(x = resultat1)
}
}
