% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_parse_pe}
\alias{inpi_parse_pe}
\title{inpi_parse_pe}
\usage{
inpi_parse_pe(x)
}
\arguments{
\item{x}{la reponse json a parser}
}
\value{
json parse de la reponse a la requete
}
\description{
retourne le json parse de reponse d'une requete sur personne exploitant
}
\examples{
\dontrun{
inpi_parse_pe(x = resultat1)
}
}
