% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_parse_pm}
\alias{inpi_parse_pm}
\title{inpi_parse_pm}
\usage{
inpi_parse_pm(x)
}
\arguments{
\item{x}{la reponse json a parser}
}
\value{
json parse de la reponse a la requete
}
\description{
retourne le json parse de reponse d'une requete sur personne morale
}
\examples{
inpi_parse_pm(x = resultat1)
}
