% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_parse_pp}
\alias{inpi_parse_pp}
\title{inpi_parse_pp}
\usage{
inpi_parse_pp(x)
}
\arguments{
\item{x}{la reponse json a parser}
}
\value{
json parse de la reponse a la requete
}
\description{
retourne le json parse de reponse d'une requete sur personne physique
}
\examples{
\dontrun{
inpi_parse_pp(x = resultat1)
}
}
