% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_pouvoirs}
\alias{inpi_pouvoirs}
\title{inpi_pouvoirs}
\usage{
inpi_pouvoirs(x)
}
\arguments{
\item{x}{la reponse json a parser}
}
\value{
data.frame avec les infos sur les pouvoirs des PM
}
\description{
retourne les colonnes du pouvoir !
}
\examples{
\dontrun{
inpi_pouvoirs(x = resultat1)
}
}
