% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_relocate}
\alias{inpi_relocate}
\title{inpi_relocate}
\usage{
inpi_relocate(x)
}
\arguments{
\item{x}{la reponse json a parser}
}
\value{
data.frame avec les colonnes d'adresse dans le bon ordre
}
\description{
retourne les colonnes de l'adresse dans le bon ordre
}
\examples{
\dontrun{
inpi_relocate(x = resultat1)
}
}
