% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_test_persoju}
\alias{inpi_test_persoju}
\title{inpi_test_persoju}
\usage{
inpi_test_persoju(x)
}
\arguments{
\item{x}{la requete a analyser}
}
\value{
vecteur de la (ou des) personnalite(s) juridique(s) de la requete
}
\description{
retourne la personnalite juridique d'une requete sur siren(s)
}
\examples{
inpi_test_persoju(x = resultat1)
}
