% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_test_type_pers}
\alias{inpi_test_type_pers}
\title{inpi_test_type_pers}
\usage{
inpi_test_type_pers(x)
}
\arguments{
\item{x}{la requete a analyser}
}
\value{
vecteur du ou des types de personne
}
\description{
retourne le type de personne : P / M / E
}
\examples{
inpi_test_type_pers(x = resultat1)
}
