% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_test_var}
\alias{inpi_test_var}
\title{inpi_test_var}
\usage{
inpi_test_var(x, var = "")
}
\arguments{
\item{x}{la variable a chercher}
}
\value{
tibble des moda et val de la var cherchee (si elle existe)
}
\description{
retourne la valeur d'une variable si elle existe, sinon un message d'erreur
}
\examples{
inpi_test_var(x)
}
