% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_vital}
\alias{inpi_vital}
\title{inpi_vital}
\usage{
inpi_vital(x)
}
\arguments{
\item{x}{la reponse json a parser}
}
\value{
data.frame avec qques colonnes
}
\description{
retourne les colonnes vitales
}
\examples{
\dontrun{
inpi_vital(x = resultat1)
}
}
