% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiinpifun.R
\name{inpi_get_by_acti}
\alias{inpi_get_by_acti}
\title{inpi_get_by_acti}
\usage{
inpi_get_by_acti(activite = "COMMERCIALE")
}
\arguments{
\item{listesiren}{un vecteur contenant les sirens a interroger}

\item{simp}{simplifier le json ? default = TRUE}
}
\value{
json non parse de la reponse a la requete sur activite
}
\description{
retourne le json de reponse de la requete sur un secteur d'activite
}
\examples{
inpi_get_by_acti(activite = "ARTISANALE", simp = TRUE)
}
